code cleanup. added comments
