+++ app/dimension/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-# import dash_bootstrap_components as dbc
 import dash_html_components as html
 
 from app.custom_widgets import custom_button
@@ -9,18 +8,11 @@ from app.dimension.simulation_widgets import coordinate_grid
 from app.dimension.simulation_widgets import environment
 from app.modal.advance_settings import advance_settings
 
-# from dash.dependencies import Input
-# from dash.dependencies import Output
-# from dash.dependencies import State
-# from dash.exceptions import PreventUpdate
-# from app.app import app
-
-
 __author__ = ["Deepansh J. Srivastava"]
 __email__ = ["deepansh2012@gmail.com"]
 
 
-# Advance settings ------------------------------------------------------------------ #
+# Advanced settings ----------------------------------------------------------------- #
 advance_setting_button = custom_button(
     icon_classname="fas fa-cog",
     id="advance_setting",
@@ -33,18 +25,31 @@ dimension_toolbar = html.Div([advance_setting_button, advance_settings])
 
 # dimension parameters
 def make_dimension(i):
+    """Create a spectral dimension interface."""
     row1 = html.Div(
         [
+            # create environment => widgets for
+            # 1) isotope,
+            # 2) magnetic flux density,
+            # 3) rotor frequency, and
+            # 4) rotor angle
             custom_collapsible(
                 text="Environment",
                 identity=f"environment_id-{i}",
                 children=environment(i),
             ),
+            # create line broadening => widgets for
+            # 1) apodization function and
+            # 2) apodization factor,
             custom_collapsible(
                 text="Line broadening",
                 identity=f"post_simulation_id-{i}",
                 children=gaussian_linebroadening_widget(i),
             ),
+            # create coordinate grid => widgets for
+            # 1) number of points,
+            # 2) spectral width, and
+            # 3) reference offset
             custom_collapsible(
                 text="Coordinate grid",
                 identity=f"coordinate_grid_id-{i}",
@@ -54,7 +59,6 @@ def make_dimension(i):
         id=f"dimension-tab-scroll-{i}",
     )
     dimension_contents = html.Div(children=[row1])
-
     return dimension_contents
 
 
@@ -65,7 +69,6 @@ dimension_body = html.Div(
         html.Div(
             [
                 html.H4("Spectral Dimension", id="dimension-card-title"),
-                # dbc.Button("+", id="add-dimension"),
                 dimension_toolbar,
             ],
             className="card-header",
@@ -77,24 +80,3 @@ dimension_body = html.Div(
 )
 
 dimension_body_card = html.Div(dimension_body, id="dimension-card-body")
-
-
-# @app.callback(
-#     [Output("dimension-tabs", "children"),
-#      Output("local-dimension-max-index", "data")],
-#     [Input("add-dimension", "n_clicks")],
-#     [State("dimension-tabs", "children"),
-#      State("local-dimension-max-index", "data")],
-# )
-# def add_dimension_tab(n, children, max_index):
-#     if max_index == 1:
-#         raise PreventUpdate
-
-#     if max_index is None:
-#         max_index = 0
-
-#     if n:
-#         children.append(make_dimension(max_index + 1))
-#         return [children, max_index + 1]
-
-#     return [children, max_index]
