+++ app/importer.py
@@ -17,11 +17,11 @@ from dash.dependencies import State
 from dash.exceptions import PreventUpdate
 from mrsimulator import Isotopomer
 
-from app.app import app
-from app.custom_widgets import custom_button
-from app.custom_widgets import label_with_help_button
-from app.custom_widgets import print_info
-from app.isotopomer.draft import get_all_isotopomer_dropdown_options
+from .app import app
+from .custom_widgets import custom_button
+from .custom_widgets import label_with_help_button
+from .isotopomer.util import get_all_isotopomer_dropdown_options
+from .isotopomer.util import print_info
 
 __author__ = "Deepansh J. Srivastava"
 __email__ = ["deepansh2012@gmail.com"]
@@ -528,7 +528,6 @@ def update_isotopomers(
 def filter_dict(dict1):
     dict_new = {}
     for key, val in dict1.items():
-        # print(key, val)
         if val is not None:
             dict_new[key] = val
         if isinstance(val, dict):
@@ -631,8 +630,7 @@ def update_dropdown_options(local_isotopomer_data, old_isotope, config):
         isotope_dropdown_options,
         isotope,
         isotopomer_dropdown_options,
-        print_info(local_isotopomer_data)
-        # json.dumps(local_isotopomer_data, indent=2),
+        print_info(local_isotopomer_data),
     ]
